a basic sphere

+++ RRR/Sphere.h
@@ -4,7 +4,7 @@
 class Sphere : public Object
 {
 public:
-	Sphere(Point o, int radius):
+	Sphere(Point o, double radius):
 		m_orgin(o),m_radius(radius)
 	{
 		m_material.m_rgb = RGB{ 0.8,0,0 };
@@ -14,7 +14,7 @@ public:
 
 public:
 	Point	m_orgin;
-	int		m_radius;
+	double		m_radius;
 
 };
